+++ colormaps/__init__.py
@@ -9,3 +9,4 @@ from __future__ import division
 from colormaps.core import GradientColormap
 from colormaps.core import DiscreteColormap
 from colormaps.core import get
+from colormaps.core import registered